verify test works once increasing datatype granularity for uint16_t

+++ src/services/IServiceA.cpp
@@ -25,11 +25,11 @@ void IServiceA::SetAll(string val) {
   all = {++id, val};
 }
 
-uint8_t IServiceA::getId(int index) {
+uint16_t IServiceA::getId(int index) {
   return elements.at(index).id;
 }
 
-uint8_t IServiceA::getAllId() {
+uint16_t IServiceA::getAllId() {
   return all.id;
 }
